pyproject.toml
[tool.poetry]
name = "memleak-example"
version = "0.1.0"
description = ""
authors = ["Sean Stewart <sean.stewart@mavenclinic.com>"]
readme = "README.md"
packages = [
    { include = "example", from = "src" },
    { include = "vendor", from = "src" }
]
[tool.poetry.scripts]
serve = "example.app:run"

[tool.poetry.dependencies]
python = "^3.12"
ddtrace = "^2.18.0"
aiohttp = "^3.11.11"
faker = "^33.1.0"
orjson = "^3.10.12"
python-dotenv = { version = "^1.0.1", extras = ["cli"] }
structlog = "^24.4.0"
uvloop = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
